Move Cas2v2 API models from /api/model to /cas2/model.
