+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2/model/LatestCas2v2StatusUpdate.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import io.swagger.v3.oas.annotations.media.Schema
