+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2/model/UpdateCas2v2Assessment.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
